feat: Ui Done and animation added; Implement multiple code changes to enhance functionality and improve performance also choice
