+++ apps/web/package.json
@@ -11,11 +11,8 @@
     "check-types": "tsc --noEmit"
   },
   "dependencies": {
-
-    "@radix-ui/react-checkbox": "^1.3.2",
-
     "@dnd-kit/core": "^6.3.1",
-
+    "@radix-ui/react-checkbox": "^1.3.2",
     "@radix-ui/react-dialog": "^1.1.14",
     "@radix-ui/react-dropdown-menu": "^2.1.15",
     "@radix-ui/react-select": "^2.2.5",
@@ -37,7 +34,8 @@
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
     "react-toastify": "^11.0.5",
-    "tailwind-merge": "^3.3.1"
+    "tailwind-merge": "^3.3.1",
+    "uuid": "^11.1.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.1",
@@ -48,6 +46,7 @@
     "@types/node": "^24.0.15",
     "@types/react": "19.1.8",
     "@types/react-dom": "19.1.6",
+    "@types/uuid": "^10.0.0",
     "@types/webpack": "5.28.5",
     "@zeit/next-css": "^1.0.1",
     "autoprefixer": "^10.4.21",
